Add files via upload
This is a userform in VBA-Excel to calculate the Age in Years, Months and Days and calculate the BMI(Body Mass Index) of a person. The Ideal BMI of Human should be between 20 and 25. The DOB(Date of Birth) is used to calculate the Age and Height and Weight is used to calculate the BMI of a person. 
Height should be in cms and Weight should be in Kgs. 
If the person, clicks on cancel, it simply closes the form. Calculate command button, calculates the Age and BMI.
There is also a activeX control button, on the sheet, for the user to access the userform.